+++ examples/conversations/service/with-notifications-and-configuration/main.tf
@@ -3,7 +3,7 @@ resource "twilio_conversations_service" "service" {
 }
 
 resource "twilio_conversations_service_configuration" "configuration" {
-  service_sid = twilio_conversations_service.service.sid
+  service_sid          = twilio_conversations_service.service.sid
   reachability_enabled = false
 }
 
@@ -13,18 +13,18 @@ resource "twilio_conversations_service_notification" "service_notification" {
   log_enabled = true
 
   new_message {
-    enabled = true
+    enabled             = true
     badge_count_enabled = true
-    template = "$${CONVERSATION}:$${PARTICIPANT}: $${MESSAGE}"
+    template            = "$${CONVERSATION}:$${PARTICIPANT}: $${MESSAGE}"
   }
 
   added_to_conversation {
-    enabled = true
+    enabled  = true
     template = "You have been added to the conversation $${CONVERSATION} by $${PARTICIPANT}"
   }
 
   removed_from_conversation {
-    enabled = true
+    enabled  = true
     template = "$${PARTICIPANT} has removed you from the conversation $${CONVERSATION}"
   }
 }