+++ examples/video/composition_hook/outputs.tf
@@ -1,0 +1,4 @@
+output "composition_hook" {
+  description = "The Generated Composition Hook"
+  value       = twilio_video_composition_hook.composition_hook
+}