+++ twilio/config.go
@@ -14,6 +14,7 @@ import (
 	studio "github.com/RJPearson94/twilio-sdk-go/service/studio/v2"
 	taskrouter "github.com/RJPearson94/twilio-sdk-go/service/taskrouter/v1"
 	trunking "github.com/RJPearson94/twilio-sdk-go/service/trunking/v1"
+	video "github.com/RJPearson94/twilio-sdk-go/service/video/v1"
 	"github.com/RJPearson94/twilio-sdk-go/session"
 	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
@@ -35,7 +36,9 @@ func (config *Config) Client() (interface{}, diag.Diagnostics) {
 	}
 
 	sess := session.New(creds)
-	sdkConfig := &client.Config{}
+	sdkConfig := &client.Config{
+		// DebugEnabled: utils.Bool(true),
+	}
 
 	client := &common.TwilioClient{
 		AccountSid:       config.AccountSid,
@@ -52,6 +55,7 @@ func (config *Config) Client() (interface{}, diag.Diagnostics) {
 		SIPTrunking:   trunking.New(sess, sdkConfig),
 		Studio:        studio.New(sess, sdkConfig),
 		TaskRouter:    taskrouter.New(sess, sdkConfig),
+		Video:         video.New(sess, sdkConfig),
 	}
 	return client, nil
 }