+++ twilio/services.go
@@ -16,6 +16,7 @@ import (
 	"github.com/RJPearson94/terraform-provider-twilio/twilio/internal/services/sip_trunking"
 	"github.com/RJPearson94/terraform-provider-twilio/twilio/internal/services/studio"
 	"github.com/RJPearson94/terraform-provider-twilio/twilio/internal/services/taskrouter"
+	"github.com/RJPearson94/terraform-provider-twilio/twilio/internal/services/video"
 	"github.com/RJPearson94/terraform-provider-twilio/twilio/internal/services/voice"
 )
 
@@ -35,6 +36,7 @@ func SupportedServices() []common.ServiceRegistration {
 		sip.Registration{},
 		sip_trunking.Registration{},
 		taskrouter.Registration{},
+		video.Registration{},
 		voice.Registration{},
 	}
 }